+++ internal/accrual/utils/database/migrations/000001_accrual_order.up.sql
@@ -2,5 +2,5 @@ CREATE TABLE accrual (
     id serial primary key,
     order_number varchar(60) unique,
     status varchar(15),
-    accrual int
+    accrual float
 );
