+++ internal/accrual/utils/database/migrations/000002_items.up.sql
@@ -2,5 +2,5 @@ CREATE TABLE items (
     id serial primary key,
     order_number varchar(60),
     description varchar(255),
-    price int
+    price float
 );
