+++ internal/accrual/utils/database/migrations/000003_goods.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE goods (
     id serial primary key,
     match varchar(60) not null unique,
-    reward int,
+    reward float,
     reward_type varchar(2)
 );
