+++ internal/gophermart/utils/types/types.go
@@ -67,7 +67,7 @@ type Order struct {
 	User       string
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    int       `json:"accrual,omitempty"`
+	Accrual    float64   `json:"accrual,omitempty"`
 	UploadedAt time.Time `json:"uploaded_at"`
 }
 
@@ -75,7 +75,7 @@ func (order Order) MarshalJSON() ([]byte, error) {
 	var tmp struct {
 		Number     string    `json:"number"`
 		Status     string    `json:"status"`
-		Accrual    int       `json:"accrual,omitempty"`
+		Accrual    float64   `json:"accrual,omitempty"`
 		UploadedAt time.Time `json:"uploaded_at"`
 	}
 	tmp.Number, tmp.Status, tmp.Accrual, tmp.UploadedAt = order.Number, order.Status, order.Accrual, order.UploadedAt
